notebooks/mi-proyecto/scripts/analytics/requirements.txt
spai>=2024.6.6
rasterio
geopandas
Pillow
scikit-image
tqdm
matplotlib